+++ runtime/src/openxla/runtime/nvgpu/cudnn/cudnn_api.h
@@ -216,13 +216,13 @@ iree::StatusOr<iree::vm::ref<CudnnTensor>> CreateTensor(
 // Creates a pointwise unary operation.
 iree::StatusOr<iree::vm::ref<CudnnTensor>> CreatePointwiseUnary(
     openxla_cudnn_dynamic_symbols_t* syms, cudnnPointwiseMode_t mode,
-    CudnnTensor& x, float alpha, int64_t uid, int64_t alignment,
+    CudnnTensor& x, double alpha, int64_t uid, int64_t alignment,
     bool is_virtual);
 
 // Creates a pointwise binary operation.
 iree::StatusOr<iree::vm::ref<CudnnTensor>> CreatePointwiseBinary(
     openxla_cudnn_dynamic_symbols_t* syms, cudnnPointwiseMode_t mode,
-    CudnnTensor& x, float alpha, CudnnTensor& b, float alpha2, int64_t uid,
+    CudnnTensor& x, double alpha, CudnnTensor& b, double alpha2, int64_t uid,
     int64_t alignment, bool is_virtual);
 
 // Creates a relu operation.
